tests/pyproject.toml
[project]
name = "mason-tests"
version = "0.0.0"
description = "eval suite for mason skills"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8",
    "pyyaml>=6",
    #
]

[tool.pytest.ini_options]
testpaths = ["."]
addopts = "-ra"
